requirements.txt
sounddevice>=0.5.2
numpy>=2.3.3
SpeechRecognition>=3.14.3
pick>=2.4.0
langchain>=0.3.27
langchain-core>=0.3.77
langchain-nvidia-ai-endpoints>=0.3.18
pydantic>=2.11.9
python-dotenv>=1.1.1
pillow>=11.3.0
customtkinter>=5.2.2
